requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.4
django-cors-headers==3.1.0
django-extensions==2.2.1
django-heroku==0.3.1
django-rest-swagger==2.1.2
django-rq==2.1.0
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
gunicorn==19.9.0
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
nltk==3.4.4
openapi-codec==1.3.2
psycopg2==2.8.3
pycodestyle==2.5.0
pydot==1.4.1
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.2
python-decouple==3.1
pytz==2019.2
redis==3.3.7
requests==2.22.0
rq==1.1.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
textblob==0.15.3
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
whitenoise==4.1.3
wrapt==1.11.2
